made metadata subdir
