--- ethsigner/signer/multiplatform/src/main/java/tech/pegasys/ethsigner/signer/multiplatform/AzureSigningMetadataFile.java
@@ -1,35 +1,0 @@
-/*
- * Copyright 2019 ConsenSys AG.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-package tech.pegasys.ethsigner.signer.multiplatform;
-
-import tech.pegasys.ethsigner.core.signing.TransactionSigner;
-import tech.pegasys.ethsigner.signer.azure.AzureConfig;
-
-class AzureSigningMetadataFile extends SigningMetadataFile {
-
-  private final AzureConfig config;
-
-  public AzureSigningMetadataFile(final String filename, final AzureConfig config) {
-    super(filename);
-    this.config = config;
-  }
-
-  public AzureConfig getConfig() {
-    return config;
-  }
-
-  @Override
-  public TransactionSigner createSigner(final MultiSignerFactory factory) {
-    return factory.createSigner(this);
-  }
-}
