+++ ethsigner/signer/multiplatform/src/main/java/tech/pegasys/ethsigner/signer/multiplatform/MultiPlatformTransactionSignerProvider.java
@@ -12,10 +12,14 @@
  */
 package tech.pegasys.ethsigner.signer.multiplatform;
 
+import tech.pegasys.ethsigner.TransactionSignerInitializationException;
 import tech.pegasys.ethsigner.core.signing.TransactionSigner;
 import tech.pegasys.ethsigner.core.signing.TransactionSignerProvider;
 import tech.pegasys.ethsigner.signer.azure.AzureKeyVaultAuthenticator;
 import tech.pegasys.ethsigner.signer.azure.AzureKeyVaultTransactionSignerFactory;
+import tech.pegasys.ethsigner.signer.filebased.FileBasedSignerFactory;
+import tech.pegasys.ethsigner.signer.multiplatform.metadata.AzureSigningMetadataFile;
+import tech.pegasys.ethsigner.signer.multiplatform.metadata.FileBasedSigningMetadataFile;
 
 import java.util.Objects;
 import java.util.Optional;
@@ -25,37 +29,64 @@ import java.util.stream.Collectors;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-public class MultiPlatformTransactionSignerProvider implements TransactionSignerProvider {
+public class MultiPlatformTransactionSignerProvider
+    implements TransactionSignerProvider, MultiSignerFactory {
 
   private static final Logger LOG = LogManager.getLogger();
 
   private final SigningMetadataTomlConfigLoader signingMetadataTomlConfigLoader;
-  private final MultiSignerFactory signerFactory;
+  final AzureKeyVaultTransactionSignerFactory azureFactory;
 
   MultiPlatformTransactionSignerProvider(
       final SigningMetadataTomlConfigLoader signingMetadataTomlConfigLoader) {
     this.signingMetadataTomlConfigLoader = signingMetadataTomlConfigLoader;
 
     final AzureKeyVaultAuthenticator azureAuthenticator = new AzureKeyVaultAuthenticator();
-    final AzureKeyVaultTransactionSignerFactory azureFactory =
-        new AzureKeyVaultTransactionSignerFactory(azureAuthenticator);
-
-    signerFactory = new MultiSignerFactory(azureFactory);
+    azureFactory = new AzureKeyVaultTransactionSignerFactory(azureAuthenticator);
   }
 
   @Override
   public Optional<TransactionSigner> getSigner(final String address) {
     return signingMetadataTomlConfigLoader
         .loadMetadataForAddress(address)
-        .map(metaDataFile -> metaDataFile.createSigner(signerFactory));
+        .map(metaDataFile -> metaDataFile.createSigner(this));
   }
 
   @Override
   public Set<String> availableAddresses() {
     return signingMetadataTomlConfigLoader.loadAvailableSigningMetadataTomlConfigs().stream()
-        .map(metaDataFile -> metaDataFile.createSigner(signerFactory))
+        .map(metaDataFile -> metaDataFile.createSigner(this))
         .filter(Objects::nonNull)
         .map(TransactionSigner::getAddress)
         .collect(Collectors.toSet());
+  }
+
+  @Override
+  public TransactionSigner createSigner(final AzureSigningMetadataFile metaData) {
+    final TransactionSigner signer = azureFactory.createSigner(metaData.getConfig());
+    final String signerAddress = signer.getAddress();
+    if (!signerAddress.equals(metaData.getBaseFilename())) {
+      LOG.error(
+          "Azure Signer's Ethereum Address ({}) does not align with metadata filename ({})",
+          signerAddress,
+          metaData.getBaseFilename());
+      throw new IllegalArgumentException("Mismatch between signer and filename.");
+    }
+    return signer;
+  }
+
+  @Override
+  public TransactionSigner createSigner(final FileBasedSigningMetadataFile signingMetadataFile) {
+    try {
+      final TransactionSigner signer =
+          FileBasedSignerFactory.createSigner(
+              signingMetadataFile.getKeyPath(), signingMetadataFile.getPasswordPath());
+      LOG.debug("Loaded signer with key '{}'", signingMetadataFile.getKeyPath().getFileName());
+      return signer;
+    } catch (final TransactionSignerInitializationException e) {
+      LOG.warn(
+          "Unable to load signer with key '{}'", signingMetadataFile.getKeyPath().getFileName(), e);
+      return null;
+    }
   }
 }
